En BattleScreen (no interactor) se mueve la pantalla si el ratón se pone en la orilla de ella

+++ Cliente/Controls/SensorialExtremos.cs
@@ -55,7 +55,7 @@ namespace KarTac.Cliente.Controls
 				v += new Vector2 (0, 1);
 
 			if (v != Vector2.Zero)
-				AlHacerPresión?.Invoke (v * gameTime.ElapsedGameTime.TotalMilliseconds);
+				AlHacerPresión?.Invoke (v * (float)gameTime.ElapsedGameTime.TotalMilliseconds);
 		}
 
 		public void LoadContent ()
